src/main/java/com/mycompany/CalculatorCli.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.mycompany;

import java.io.BufferedReader;
import java.io.InputStreamReader;
import java.io.Reader;
import java.util.function.DoubleConsumer;
import java.util.stream.Stream;

/**
 *
 * @author user
 */
public class CalculatorCli {
    private final Calculator calculator;

    public CalculatorCli(Calculator claculator) {
        this.calculator = claculator;
    }
    
    public void runInteractiveSession(Reader reader) {
        runInteractiveSession(reader, System.out::println);
    }
    
    public void runInteractiveSession(Reader reader, DoubleConsumer resultConsumer) {
        new BufferedReader(reader).lines()
                    .flatMap(s -> Stream.of(s.split(";")))
                    .filter(s -> !s.trim().isEmpty())
                    .mapToDouble(calculator::calculate)
                    .forEach(resultConsumer);
                
    }
    
        public static void main(String[] args) throws Exception {
        CalculatorCli calculatorCli = new CalculatorCli(new CalculatorImp());
        try (Reader reader = new InputStreamReader(System.in)) {
            calculatorCli.runInteractiveSession(reader);
        }
    }
}

